Temperature Readings from Climacell

This function will make a call to the Climacell API and retrieve temperature related variables.

climacell_temperature(
  api_key,
  lat,
  long,
  timestep,
  start_time = NULL,
  end_time = NULL
)

Arguments

api_key
character string representing the private API key. Provided by user or loaded automatically from environment variable (environment variable must be called "CLIMACELL_API")

lat
a numeric value (or a string that can be coerced to numeric) representing the latitude of the location

long
a numeric value (or a string that can be coerced to numeric) representing the longitude of the location

timestep
a 'step' value for the time. Choose one of the following valid values: c('1d', '1h', '30m','15m','5m','1m','current')

start_time
the start date and time of the query in ISO8601 format. This value cannot be more than 6 hours prior to the current local time. OPTIONAL if timestep is 'current'.

end_time
the end date and time of the query in ISO8601 format. The maximum end date/time depends on the timestep value chosen (see vignette for more information). OPTIONAL if timestep is 'current' or you wish to get the maximum results possible.

Value

a tibble

Details

climacell_temperature returns a tibble that consists of temperature related variables (returned values are in metric units) using the Climacell API. These variables consist of temperature, a "feels like" temperature, dewpoint, and humidity.

Examples

if (FALSE) {
climacell_temperature(
  api_key = Sys.getenv('CLIMACELL_API'),
  lat = 0,
  long = 0,
  timestep = 'current')
}